Do the step "Iteration 2: Recipe Page" form Odin
Create a new directory within the odin-recipes directory and name it
recipes.
Create a new HTML file within the recipes directory and name it after
the recipe it will contain. For example lasagna.html. You can use the
name of your favorite dish or, if you need some inspiration, you can
find a recipe to use here.
For now, just include an h1 heading with the recipe’s name as its
content.
Back in the index.html file, add a link to the recipe page you just
created. Example: Under the <h1>Odin Recipes</h1> heading, write out
the link like so: <a href="recipes/recipename.html">Recipe Title</a>.
The text of the link should again be the recipe name.
I did it for three recipes already.